# src/pwa/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebAZ Protocol API",
    "version": "0.4.14",
    "description": "Auto-generated endpoint inventory (640 endpoints, 17 with full schema). See docs/ for design context."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "api_key"
      },
      "webauthnToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-WebAuthn-Token"
      }
    }
  },
  "paths": {
...
    "/api/flash-sales/:id": {
      "delete": {
        "summary": "取消（仅 seller 自己，且未开始）",
        "tags": [],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/products/:product_id/flash-sale": {
      "get": {
        "summary": "公开：商品当前生效的 flash sale",
        "tags": [],
        "security": []
      },
      "post": {
        "summary": "POST /api/products/:product_id/flash-sale",
        "tags": [],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/sellers/me/flash-sales": {
      "get": {
        "summary": "seller 自己的 flash sales（全部状态）",
        "tags": [],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}